fix: 全ブランチレビューの機械的指摘を是正(未使用export/死にAPI/旧仕様文言/D&D誤clear等)
