+++ moorestech_client/Assets/Scripts/Client.Game/InGame/UI/UIState/State/SkitState.cs
@@ -28,7 +28,7 @@ namespace Client.Game.InGame.UI.UIState.State
                 _playerInventoryViewController.SetActive(false);
             }
 
-            // GameStateControllerでスキット状態に遷移（ホットバーの非表示を含む）
+            // GameStateControllerでスキット状態に遷移
             GameStateController.ChangeState(GameStateType.Skit);
 
             KeyControlDescription.Instance.SetText("");
