+++ moorestech_server/Assets/Scripts/Core.Inventory/OpenableInventoryItemDataStoreService.cs
@@ -138,20 +138,7 @@ namespace Core.Inventory
         {
             return InventoryInsertItem.InsertItem(itemStacks, _inventory, _itemStackFactory, _option, InvokeEvent);
         }
-        
-        /// <summary>
-        ///     もしそのスロットに入れられるのであれば、まずはそれらのスロットに入れてから、余ったものを返す
-        /// </summary>
-        public IItemStack InsertItemWithPrioritySlot(IItemStack itemStack, int[] prioritySlots)
-        {
-            return InventoryInsertItem.InsertItemWithPrioritySlot(itemStack, _inventory, _itemStackFactory, _option, prioritySlots, InvokeEvent);
-        }
-        
-        public IItemStack InsertItemWithPrioritySlot(ItemId itemId, int count, int[] prioritySlots)
-        {
-            return InsertItemWithPrioritySlot(_itemStackFactory.Create(itemId, count), prioritySlots);
-        }
-        
+
         #endregion
     }
