+++ moorestech_web/webui/src/features/hotbar/index.ts
@@ -1,4 +1,2 @@
 export { default as HotbarPanel } from "./HotbarPanel";
 export { useHotbarDragSource } from "./useHotbarDragSource";
-export { resolveDropAction } from "./hotbarDnd";
-export type { DragEndpoint, HotbarDropAction } from "./hotbarDnd";
