+++ moorestech_web/webui/src/features/inventory/InventoryPanel/index.tsx
@@ -22,7 +22,7 @@ export default function InventoryPanel() {
 
   return (
     <GamePanel gridArea="inv" title={t(L.ui.inventory.title)} style={{ justifySelf: "start", alignSelf: "start", width: 378, minHeight: 452.391, transform: "translate(0.783px, 0.783px)", "--panel-left": "-2.22px", "--panel-right": "-2.22px", "--title-shift-x": "-1.96px", "--title-scale-x": 0.919, "--title-scale-y": 0.924 } as CSSProperties}>
-      <SlotGrid testId="main-grid" cols={6} style={GRID_STYLE}>
+      <SlotGrid testId="main-grid" cols={9} style={GRID_STYLE}>
         {inventory.mainSlots.map((slot, i) => {
           const ref: SlotRef = { area: "main", slot: i };
           return (
